Rename ProjectsFolderBrowser.cpp/h to ProjectBrowser
