+++ Makefile
@@ -83,7 +83,7 @@ SRCS += src/ui/GenioWindow.cpp
 SRCS += src/ui/GoToLineWindow.cpp
 SRCS += src/ui/IconCache.cpp
 SRCS += src/ui/ProblemsPanel.cpp
-SRCS += src/ui/ProjectsFolderBrowser.cpp
+SRCS += src/ui/ProjectBrowser.cpp
 SRCS += src/ui/SearchResultPanel.cpp
 SRCS += src/ui/StyledItem.cpp
 SRCS += src/ui/ToolBar.cpp
