+++ src/git/SourceControlPanel.cpp
@@ -29,8 +29,8 @@
 #include "GitAlert.h"
 #include "GTextAlert.h"
 #include "Log.h"
+#include "ProjectBrowser.h"
 #include "ProjectFolder.h"
-#include "ProjectsFolderBrowser.h"
 #include "RepositoryView.h"
 #include "StringFormatter.h"
 #include "Utils.h"
