+++ src/ui/GenioWindow.cpp
@@ -59,9 +59,9 @@
 #include "Log.h"
 #include "LSPEditorWrapper.h"
 #include "ProblemsPanel.h"
+#include "ProjectBrowser.h"
 #include "ProjectFolder.h"
 #include "ProjectItem.h"
-#include "ProjectsFolderBrowser.h"
 #include "QuitAlert.h"
 #include "RemoteProjectWindow.h"
 #include "SearchResultPanel.h"
